+++ include/depthai-shared/datatype/RawImgFrame.hpp
@@ -2,7 +2,7 @@
 
 #include <unordered_map>
 
-#include "depthai-shared/common/ImgTransformation.hpp"
+#include "depthai-shared/common/ImgTransformations.hpp"
 #include "depthai-shared/common/Timestamp.hpp"
 #include "depthai-shared/datatype/RawBuffer.hpp"
 #include "depthai-shared/utility/Serialization.hpp"
